ai-trial-game/backend/requirements.txt
# Web framework
fastapi>=0.100.0
uvicorn>=0.23.0

# LLM
anthropic>=0.18.0
openai>=1.0.0

# SpoonOS (local editable install)
-e ../../spoon-core-main/spoon-core-main

# Blockchain integration
web3>=6.0.0

# Utilities
python-dotenv>=1.0.0
pydantic>=2.0.0
